fixed some compiler warnings

+++ src/geometry/PolySetUtils.cc
@@ -96,7 +96,7 @@ std::unique_ptr<PolySet> tessellate_faces(const PolySet& polyset)
   std::vector<Vector3f> verts;
   std::vector<int> indexMap(polyset.vertices.size());
   verts.reserve(polyset.vertices.size());
-  for (int i = 0; i < polyset.vertices.size(); ++i) {
+  for (size_t i = 0; i < polyset.vertices.size(); ++i) {
     if (used[i]) {
       indexMap[i] = verts.size();
       verts.emplace_back(polyset.vertices[i].cast<float>());
